etl/lending_club/requirements.txt
requests==2.31.0
pandas==1.5.3
numpy==1.24.4
google-cloud-storage==2.9.0
functions-framework==3.7.0
Flask==2.3.2
pyarrow==12.0.0